+++ test/unit/org/apache/cassandra/cql3/CQLTester.java
@@ -1528,6 +1528,9 @@ public abstract class CQLTester
         if (value instanceof Float)
             return FloatType.instance;
 
+        if (value instanceof Duration)
+            return DurationType.instance;
+
         if (value instanceof Double)
             return DoubleType.instance;